requirements.txt
anyio==4.4.0
attrs==23.2.0
Automat==22.10.0
backports.tarfile==1.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
exceptiongroup==1.2.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.2.0
incremental==24.7.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jeepney==0.8.0
joblib==1.4.2
keyring==25.2.1
llvmlite==0.43.0
more-itertools==10.3.0
numba==0.60.0
numpy==1.25.2
pandas==2.2.2
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-okx==0.3.1
pytz==2024.1
requests==2.32.3
schedule==1.2.2
SecretStorage==3.3.3
six==1.16.0
sniffio==1.3.1
TA-Lib==0.4.32
toml==0.10.2
tomli==2.0.1
Twisted==24.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zipp==3.19.2
zope.interface==6.4.post2
